+++ src/objects/impls/nuke.rs
@@ -1,8 +1,8 @@
-use crate::{macros::*, objects::Nuke};
+use crate::{local::RoomName, macros::*, objects::Nuke};
 
 simple_accessors! {
     Nuke;
     // id from HasID
-    (lauch_room_name -> lauchRoomName -> String),
+    (lauch_room_name -> lauchRoomName -> RoomName),
     (time_to_land -> timeToLand -> u32),
 }